edit few things

+++ database/migrations/2025_01_16_090858_create_scores_table.php
@@ -12,15 +12,17 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('scores', function (Blueprint $table) {
-            $table->id(); // Adds an auto-incrementing primary key column (id)
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            // Foreign key linking to the users table.
-            // ->constrained() automatically sets up the foreign key relationship.
-            // ->onDelete('cascade') ensures that if a user is deleted, their related scores will also be deleted.
-            $table->integer('score'); // User's score
-            $table->timestamps(); // Record when the quiz was taken
-        });
+        if (!Schema::hasTable('scores')) {
+            Schema::create('scores', function (Blueprint $table) {
+                $table->id(); // Adds an auto-incrementing primary key column (id)
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                // Foreign key linking to the users table.
+                // ->constrained() automatically sets up the foreign key relationship.
+                // ->onDelete('cascade') ensures that if a user is deleted, their related scores will also be deleted.
+                $table->integer('score'); // User's score
+                $table->timestamps(); // Record when the quiz was taken
+            });
+        }
     }
 
     /**
